moved the guided tutorial to the map page
